test(maru): stabilize app integration tests under constrained CI runners
Cap the integrationTest fork heap (1.5 GB + G1) so a GC pause can't be
CFS-throttled into a multi-second stop-the-world that freezes Maru's 1 s
slot timer (the root of the flaky 2 s block-time gaps observed on CI).
Scale maxParallelForks with cores (~2 cores/fork, capped at 8). Split
MaruFollowerTest into a shared base + import-blocks + syncing classes so
the suite makespan isn't bound by one ~8.5 min class. Run unit and
integration tests in a single CI step.
