+++ .github/workflows/maru-testing.yml
@@ -40,12 +40,9 @@ jobs:
           submodules: false
       - name: Setup Java and Gradle
         uses: ./.github/actions/setup-java-and-gradle
-      - name: Build maru and Unit tests
+      - name: Build, unit and integration tests
         run: |
-          ./gradlew -V maru:app:buildNeeded
-      - name: Run integration tests
-        run: |
-          ./gradlew maru:app:integrationTest
+          ./gradlew -V maru:app:buildNeeded maru:app:integrationTest
       - name: Store reports
         if: failure()
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
